Separate _select_elite_population() from `NSGAIISampler`
